fix: refactor code and database call

+++ src/modules/users/users.service.ts
@@ -11,48 +11,30 @@ export class UsersService {
     @InjectRepository(User) private readonly userRepository: Repository<User>,
   ) {}
   async create(createUserDto: CreateUserDto): Promise<User> {
-    try {
-      const user = await this.userRepository.save(createUserDto)
-      return user
-    } catch (error) {
-      throw new HttpException(error.message, 400)
-    }
+    const user = await this.userRepository.findOneBy({
+      email: createUserDto.email,
+    })
+    if (user) throw new HttpException('El usuario ya existe', 400)
+    return await this.userRepository.save(createUserDto)
   }
 
   async deleteById(id: number): Promise<User> {
-    try {
-      const user = await this.findById(id)
-      await this.userRepository.remove(user)
-      return user
-    } catch (error) {
-      throw new HttpException(error.message, 400)
-    }
+    const user = await this.userRepository.findOneBy({ id })
+    if (!user) throw new HttpException('Usuario no encontrado', 404)
+    return await this.userRepository.remove(user)
   }
 
   async findAll(): Promise<User[]> {
-    try {
-      const users = await this.userRepository.find()
-      return users
-    } catch (error) {
-      throw new HttpException(error.message, 400)
-    }
+    return await this.userRepository.find()
   }
 
-  async findById(id: number): Promise<User> {
-    try {
-      const user = await this.userRepository.find({ where: { id } })
-      return user[0]
-    } catch (error) {
-      throw new HttpException(error.message, 400)
-    }
+  async findById(id: number): Promise<User | {}> {
+    const user = await this.userRepository.findOneBy({ id })
+    return user ? user : { statusCode: 404, message: 'User not found' }
   }
 
-  async findByEmail(email: string): Promise<User> {
-    try {
-      const user = await this.userRepository.find({ where: { email } })
-      return user[0]
-    } catch (error) {
-      throw new HttpException(error.message, 400)
-    }
+  async findByEmail(email: string): Promise<User | {}> {
+    const user = await this.userRepository.findOneBy({ email })
+    return user || { statusCode: 404, message: 'User not found' }
   }
 }
